1. export through index files 2. scss for styles 3. useCallback 4. custom hooks for keeping the compnent clean 5. rename as tasks and todos or todolist and todolists
