--- src/components/Input/Input.css
@@ -1,8 +1,0 @@
-.input {
-  margin: 20px 0;
-}
-
-.input > input {
-  width: 60%;
-  margin-right: 20px;
-}
