+++ src/components/Input/Input.js
@@ -1,10 +1,14 @@
-import "./Input.css";
+import "./input.scss";
 
-const Input = ({add, input, setInput}) => {
-
+const Input = ({ add, input, setInput }) => {
   return (
     <div className="input">
-      <input type="text" value={input} onChange={(e) => setInput(e.target.value)} onKeyDown={(e) => e.key === "Enter" ? add(input) : ""}/>
+      <input
+        type="text"
+        value={input}
+        onChange={(e) => setInput(e.target.value)}
+        onKeyDown={(e) => (e.key === "Enter" ? add(input) : "")}
+      />
       <button onClick={add}>Add New</button>
     </div>
   );
