+++ src/components/Input/input.scss
@@ -1,0 +1,7 @@
+.input {
+  margin: 20px 0;
+  input {
+    width: 60%;
+    margin-right: 20px;
+  }
+}
