--- src/components/Items/Items.css
@@ -1,10 +1,0 @@
-.items {
-  margin: 20px 0;
-}
-.item {
-  margin-bottom: 5px;
-}
-.item-name {
-  margin-left: 5px;
-  margin-right: 20px;
-}
