+++ src/components/Tasks/index.js
@@ -1,0 +1,3 @@
+import Tasks from "./Tasks";
+
+export default Tasks;
