+++ src/components/Tasks/tasks.scss
@@ -1,0 +1,10 @@
+.items {
+  margin: 20px 0;
+  .item {
+    margin-bottom: 5px;
+    .item-name {
+      margin-left: 5px;
+      margin-right: 20px;
+    }
+  }
+}
