+++ src/components/Todo/todo.scss
@@ -1,0 +1,6 @@
+.actions {
+  margin: 20px 0;
+  button {
+    margin-right: 20px;
+  }
+}
