--- src/components/TodoItem/TodoItem.css
@@ -1,7 +1,0 @@
-.actions {
-  margin: 20px 0;
-}
-
-.actions > button {
-  margin-right: 20px;
-}
